Create library and tests, modularize gem.
Now the code resides in lib/gn.rb and it can be tested with
test/gn.rb. The edit-load-generate strategy was streamlined and now
the only available location for templates is ~/.gn/<plan>. This
simplifies the code, and also paves the way for commands like `gn
list`, `gn install`, `gn uninstall`, etc.